+++ lib/feature/my_job/view/posting_create_page.dart
@@ -3,10 +3,10 @@ import 'package:alter/common/util/%08formater/formatter.dart';
 import 'package:alter/common/util/%08formater/number_with_comma_formatter.dart';
 import 'package:alter/common/util/logger.dart';
 import 'package:alter/feature/home/model/posting_response_model.dart';
-import 'package:alter/feature/home/view/widget/job_dropdown_widget.dart';
-import 'package:alter/feature/home/view/widget/paytype_selector.dart';
-import 'package:alter/feature/home/view/widget/schedule_widget.dart';
-import 'package:alter/feature/home/view_model/posting_create_view_model.dart';
+import 'package:alter/feature/my_job/view/widget/job_dropdown_widget.dart';
+import 'package:alter/feature/my_job/view/widget/paytype_selector.dart';
+import 'package:alter/feature/my_job/view/widget/schedule_widget.dart';
+import 'package:alter/feature/my_job/view_model/posting_create_view_model.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
